Apparently yosemite has a limit to the number of symbolic links in a directory.  This may be a pyrrhic victory.

+++ openssl/util/point.sh
@@ -1,7 +1,12 @@
 #!/bin/sh
 
-rm -f "$2"
-if test "$OSTYPE" = msdosdjgpp || test "x$PLATFORM" = xmingw ; then
+#OSX filesystem only pretends to be case insensitive.  
+os=`echo $OSTYPE | cut -c 1-6`
+if [ $os != darwin ] || [ `echo $1 | tr [:upper:] [:lower:]`  !=  `echo $2 | tr [:upper:] [:lower:]` ]; then
+	rm -f "$2"
+fi
+
+if test "$OSTYPE" = msdosdjgpp || test "x$PLATFORM" = xmingw || [ $os = darwin ] ; then
     cp "$1" "$2"
 else
     ln -s "$1" "$2"
